move ModelMetaclass to _internal._model_construction
